+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## next
+
+- CLI: show Avram specification number and result
+
 ## 0.4.1 - 2023-12-22
 
 - Update validation options to support all but externalRule